+++ lib/vagrant/plugin/manager.rb
@@ -72,13 +72,6 @@ module Vagrant
           install_lambda.call
         end
 
-        # If the version constraint is just a specific version, don't
-        # store the constraint. However, if it is a prerelease version,
-        # we DO store the constraint to avoid Bundler updating it.
-        if opts[:version] && opts[:version] =~ /^\d/ && opts[:version] !~ /[a-z]/i
-          opts.delete(:version)
-        end
-
         # Add the plugin to the state file
         @user_file.add_plugin(
           result.name,